feat: 中継する際に `adapter :socat` で socat を間にかます refactor: クラスごとの lib/ 配下のファイルに切り出した
